Move all saml protocol models into samlprotocol
